♻️ sample command api 를 effect 적용
