+++ apps/app/src/module/sample/application/port/in/SampleCommandUseCase.ts
@@ -1,9 +1,9 @@
+import type { T } from '@app/custom/effect';
 import type { DBError } from '@app/domain/error/DBError';
-import type { TaskEither } from 'fp-ts/TaskEither';
 
 import type { CreateSampleCommand } from './dto/CreateSampleCommand';
 import type { Sample } from '../../../domain/Sample';
 
 export abstract class SampleCommandUseCase {
-  abstract create(command: CreateSampleCommand): TaskEither<DBError, Sample>;
+  abstract create(command: CreateSampleCommand): T.IO<DBError, Sample>;
 }
